move specs to bottom-tabs dir, add separate types files
